fix: 🐛 improve code

+++ components/cartComponent.js
@@ -1,16 +1,26 @@
 const apiUrl = "https://ec-course-api.hexschool.io/v2";
 const apiPath = "jiayu";
+
 export default {
   template: `#cartComponent`,
   props: ["cart", "getCarts"],
   data() {
     return {
       loadingItem: "",
+      timer: 0
     };
   },
   methods: {
     async delCart(id) {
       try {
+        const swal = await Swal.fire({
+          title: "確定要刪除嗎？",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonText: "確定",
+          cancelButtonText: "先不要",
+        })
+        if(swal.isConfirmed){
         this.loadingItem = id;
 
         const url = `${apiUrl}/api/${apiPath}/cart/${id}`;
@@ -19,6 +29,7 @@ export default {
         this.getCarts();
 
         this.loadingItem = "";
+        }
       } catch (error) {
         console.log(error);
         alert(error.data.message);
@@ -26,11 +37,20 @@ export default {
     },
     async delAllCart() {
       try {
-        const url = `${apiUrl}/api/${apiPath}/carts`;
-        const { data } = await axios.delete(url);
-        alert(data.message);
-
-        this.getCarts();
+        const swal = await Swal.fire({
+          title: "確定要清空購物車嗎？",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonText: "確定",
+          cancelButtonText: "先不要",
+        })
+        if(swal.isConfirmed){
+          const url = `${apiUrl}/api/${apiPath}/carts`;
+          const { data } = await axios.delete(url);
+          alert(data.message);
+  
+          this.getCarts();
+        }
       } catch (error) {
         console.log(error);
         alert(error.data.message);
@@ -54,5 +74,11 @@ export default {
         alert(error.data.message);
       }
     },
+    debouncePutCart(item){
+      clearTimeout(this.timer);
+      this.timer = setTimeout(()=>{
+        this.putCart(item);
+      },300)
+    }
   },
 };
